refactor(api): split trpc routers/account by procedure family
Bean: api-ot6n
